requirements.txt
pandas
matplotlib
textblob
praw
python-dotenv   
